+++ test/PHPUnit/module/actes-generique/action/TedetisRecupTest.php
@@ -62,7 +62,7 @@ class TedetisRecupTest extends PastellTestCase
             'numero_de_lacte' => '201812101049',
         ]);
 
-        $documentSQL = $this->getObjectInstancier()->getInstance(Document::class);
+        $documentSQL = $this->getObjectInstancier()->getInstance(DocumentSQL::class);
         $documentSQL->setTitre($id_d, "achat d'un bus logiciel");
 
         $donneesFormulaire->setData('tedetis_transaction_id', 42);
@@ -114,11 +114,87 @@ class TedetisRecupTest extends PastellTestCase
         );
 
         $this->assertEquals(
-            [0 => 'AT_4-tampon.pdf'],
+            [0 => 'ma_premiere_annexe-tampon.pdf'],
             $donneesFormulaire->get('annexes_tamponnees')
         );
 
         $this->assertEquals("2017-12-27", $donneesFormulaire->get('date_ar'));
+    }
+
+    /**
+     * @throws Exception
+     */
+    public function testErreurAnnexes()
+    {
+        // Préparation du test
+        $curlWrapper = $this->createMock(CurlWrapper::class);
+        $curlWrapper->method("get")->willReturnCallback(function ($url) {
+
+            if ($url == "/admin/users/api-list-login.php") {
+                return true;
+            } elseif ($url == "/modules/actes/actes_transac_get_status.php?transaction=42") {
+                return "OK\n4\n" . file_get_contents(__DIR__ . "/../fixtures/aractes.xml");
+            } elseif ($url == "/modules/actes/actes_create_pdf.php?trans_id=42") {
+                return "bordereau content";
+            } elseif ($url == "/modules/actes/actes_transac_get_files_list.php?transaction=42") {
+                return file_get_contents(__DIR__ . "/../fixtures/actes_transac_get_files_list.json");
+            } elseif ($url == "/modules/actes/actes_download_file.php?file=3968&tampon=true") {
+                return "some pdf stuff tamponne";
+            } elseif ($url == "/modules/actes/actes_download_file.php?file=3969&tampon=true") {
+                return "some annexe tamponne";
+            }
+
+            throw new Exception("$url inatendu");
+        });
+
+        $curlWrapperFactory = $this->createMock(CurlWrapperFactory::class);
+        $curlWrapperFactory->method("getInstance")->willReturn($curlWrapper);
+
+        $this->getObjectInstancier()->setInstance(CurlWrapperFactory::class, $curlWrapperFactory);
+
+
+        $result = $this->getInternalAPI()->post(
+            "/entite/1/connecteur/",
+            ['libelle' => 's2low','id_connecteur' => 's2low']
+        );
+        $id_ce = $result['id_ce'];
+
+        $this->getInternalAPI()->post(
+            "/entite/1/flux/actes-generique/connecteur/$id_ce",
+            ['type' => 'TdT']
+        );
+
+        $result = $this->getInternalAPI()->post(
+            "/Document/1",
+            array('type' => 'actes-generique')
+        );
+        $id_d = $result['id_d'];
+
+
+        $this->getObjectInstancier()->getInstance(DonneesFormulaireFactory::class);
+
+        $donneesFormulaire = $this->getDonneesFormulaireFactory()->get($id_d);
+
+        $donneesFormulaire->setTabData([
+            'objet' => "achat d'un bus logiciel",
+            'numero_de_lacte' => '201812101049',
+        ]);
+
+        $documentSQL = $this->getObjectInstancier()->getInstance(DocumentSQL::class);
+        $documentSQL->setTitre($id_d, "achat d'un bus logiciel");
+
+        $donneesFormulaire->setData('tedetis_transaction_id', 42);
+        $donneesFormulaire->addFileFromData('arrete', 'mon_acte.pdf', '');
+        $donneesFormulaire->addFileFromData('autre_document_attache', 'ma_premiere_annexe_envoyée.pdf', '');
+
+        $actionChange = $this->getObjectInstancier()->getInstance(ActionChange::class);
+        $actionChange->addAction($id_d, PastellTestCase::ID_E_COL, 0, 'send-tdt', 'phpunit');
+
+        $this->expectException("Exception");
+        $this->expectExceptionMessage("Une erreur est survenue lors de la récupération des annexes tamponnées de S²low L'annexe tamponée ma_premiere_annexe.pdf ne correspond pas avec ma_premiere_annexe_envoy__e.pdf");
+        $this->getInternalAPI()->post(
+            "/entite/" . PastellTestCase::ID_E_COL . "/document/{$id_d}/action/verif-tdt"
+        );
     }
 
     /**